UI: session details and side menu [WIP]

+++ src/views/partials/side_menu.html.tpl
@@ -8,11 +8,29 @@
 {% endif %}
 <div class="side-content">
     {% if isHomepage %}
-        <div id="homepage-tab" class="homepage-tab">
-            {# TODO: #}
-            {# show last finnished project #}
-            {# total iterations implemented #}
-            {# or info about selected project #}
+        <div id="homepage-tab" class="tab homepage-tab">
+            <div class="header">
+                <h1 class="side-title">Kart Factory</h1>
+                <h2 class="side-subtitle"><span class="label">Program</span></h2>
+            </div>
+            <div class="preview-session info hidden">
+                <p class="preview-header">Session Preview</p>
+                <div class="group-number">
+                    <p class="info">Session #<span class="value number"></span></p>
+                </div>
+                <div class="group-status">
+                    <p class="info">Status: <span class="value status"></span></p>
+                </div>
+                <div class="group-stations">
+                    <p class="info">Stations: <span class="value num-stations"></span></p>
+                </div>
+                <div class="group-runs">
+                    <p class="info">Runs: <span class="value num-runs"></span></p>
+                </div>
+                <div class="group-target">
+                    <p class="info">Target: <span class="value target"></span></p>
+                </div>
+            </div>
         </div>
     {% elif title in ("Session Details") %}
         <div id="session-details-tab" class="tab session-details-tab">
@@ -20,10 +38,9 @@
                 <h1 class="side-title">Session <span class="hash">#</span>{{ session.number }}</h1>
             </div>
             <div class="buttons button-actions">
-                <a class="button button-blue button-session button-charts" href="/projects/{{ session.number }}/charts">Stats</a>
+                <a class="button button-blue button-session button-charts" href="/projects/{{ session.number }}/stats">Stats</a>
             </div>
         </div>
-        {# TODO: if Active Run / Run Details #}
     {% elif title in ("Run Details") %}
         {% set completed_run = run.status in ('FINISHED') %}
         <div id="active-run-tab" class="tab active-run-tab">
@@ -31,12 +48,18 @@
                 <h1 class="side-title">Session <span class="hash">#</span>{{ session.number }}</h1>
                 <h2 class="side-subtitle"><span class="label">Run</span><span class="value">{{ run.number }}</span></h2>
             </div>
-            <div class="buttons button-actions button-actions-run">
-                <a class="button button-blue button-start {% if completed_run %}disabled{% endif%}" href="/projects/{{ session.number }}/runs/{{ run.number }}">Start</a>
-                <a class="button button-blue button-move disabled" href="/projects/{{ session.number }}/runs/{{ run.number }}">Move</a>
-                <a class="button button-blue button-continue disabled" href="/projects/{{ session.number }}/runs/{{ run.number }}">Continue</a>
-                <a class="button button-blue button-kill disabled" href="/projects/{{ session.number }}/runs/{{ run.number }}">Kill</a>
-            </div>
+            {% if not completed_run %}
+                <div class="buttons button-actions button-actions-run">
+                    <a class="button button-blue button-start {% if completed_run %}disabled{% endif%}" href="/projects/{{ session.number }}/runs/{{ run.number }}">Start</a>
+                    <a class="button button-blue button-move disabled" href="/projects/{{ session.number }}/runs/{{ run.number }}">Move</a>
+                    <a class="button button-blue button-continue disabled" href="/projects/{{ session.number }}/runs/{{ run.number }}">Continue</a>
+                    <a class="button button-blue button-kill disabled" href="/projects/{{ session.number }}/runs/{{ run.number }}">Kill</a>
+                </div>
+            {% else %}
+                <div class="buttons button-actions">
+                    <a class="button button-blue button-run button-charts" href="/projects/{{ session.number }}/runs/{{ run.number }}/stats">Stats</a>
+                </div>
+            {% endif %}
         </div>
     {% endif %}
 </div>
